requirements.txt
asn1crypto==0.24.0
blinker==1.4
boto==2.49.0
boto3==1.9.142
botocore==1.12.142
bz2file==0.98
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
docutils==0.14
enum34==1.1.6
Flask==1.0.2
Flask-Mail==0.9.1
futures==3.2.0
gensim==3.7.2
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.3.3
MarkupSafe==1.1.1
numpy==1.16.3
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-docx==0.8.10
python-Levenshtein==0.12.0
pytz==2019.1
redis==3.3.8
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
sentry-sdk==0.7.14
six==1.12.0
smart-open==1.8.3
stripe==2.27.0
urllib3==1.25.2
Werkzeug==0.15.2
xgboost==0.82
